requirements.txt
django
djangorestframework
gunicorn
pillow
boto3
django-storages
python-decouple